Change cvssScore to float, because score has the following format x.x

+++ sonar-dependency-check-plugin/src/main/java/org/sonar/dependencycheck/base/DependencyCheckUtils.java
@@ -39,11 +39,11 @@ public final class DependencyCheckUtils {
         return new SMInputFactory(xmlFactory);
     }
 
-    public static Severity cvssToSonarQubeSeverity(String cvssScore, Double critical, Double major) {
-        double score = Double.parseDouble(cvssScore);
-        if (critical.doubleValue() > 0 && score >= critical.doubleValue()) {
+    public static Severity cvssToSonarQubeSeverity(String cvssScore, Float critical, Float major) {
+        float score = Float.parseFloat(cvssScore);
+        if (critical.floatValue() > 0 && score >= critical.floatValue()) {
             return Severity.CRITICAL;
-        } else if (major.doubleValue() > 0 && score >= major.doubleValue()) {
+        } else if (major.floatValue() > 0 && score >= major.floatValue()) {
             return Severity.MAJOR;
         } else {
             return Severity.MINOR;
